move XmlUtils to archive module
