+++ ellinet-sync/src/main/java/de/nrw/hbz/regal/sync/ingest/EllinetDigitalEntityBuilder.java
@@ -29,7 +29,7 @@ import javax.xml.xpath.XPathFactory;
 import org.w3c.dom.Element;
 import org.w3c.dom.NodeList;
 
-import de.nrw.hbz.regal.api.helper.XmlUtils;
+import de.nrw.hbz.regal.fedora.XmlUtils;
 import de.nrw.hbz.regal.sync.extern.DigitalEntity;
 import de.nrw.hbz.regal.sync.extern.StreamType;
