+++ regal-tests/src/test/java/de/nrw/hbz/regal/sync/TestEdowebDigitalEntityBuilder.java
@@ -2,7 +2,7 @@ package de.nrw.hbz.regal.sync;
 
 import org.junit.Test;
 
-import de.nrw.hbz.regal.api.helper.XmlUtils;
+import de.nrw.hbz.regal.fedora.XmlUtils;
 import de.nrw.hbz.regal.sync.extern.DigitalEntity;
 import de.nrw.hbz.regal.sync.extern.StreamType;
 import de.nrw.hbz.regal.sync.ingest.DigitoolDownloader;
